+++ CHANGELOG.md
@@ -26,6 +26,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ---
 
+## [0.1.0] - 2025-06-16 - Vlad
+
+### Updated
+
+- DB Models & Repositories
+
 ## [0.1.0] - 2025-06-15 - Vlad
 
 ### Added